Worker code clean up and refactoring
The refactoring aims at correcting dependencies in a code that creates
and sets up worker tasks, so that task initialization was happening
completelly within the implementation file of wbase/Task.cc rather
than being spread between wbase/Task.cc and wcontrol/Foreman.cc.
Altogether this is meant to improve the observability of the code and
make it easier to maintain.

Also, some minor code cleanup and removal of unused header includes
and forward declarations was made. Thould reduce the compilation
time of the relevant modules.